+++ Civi/ActionProvider/Action/Contact/Actions.php
@@ -134,6 +134,10 @@ class Actions {
       AbstractAction::SINGLE_CONTACT_ACTION_TAG,
       AbstractAction::DATA_MANIPULATION_TAG,
     ));
+    $provider->addAction('UpdateContactProfilePhoto', '\Civi\ActionProvider\Action\Contact\UpdateContactProfilePhoto', E::ts('Contact: Update Profile Photo'), array(
+      AbstractAction::SINGLE_CONTACT_ACTION_TAG,
+      AbstractAction::DATA_MANIPULATION_TAG,
+    ));
     $provider->addAction('SetCommunicationStyle', '\Civi\ActionProvider\Action\Contact\SetCommunicationStyle', E::ts("Contact: Edit communication styles"), array(
       AbstractAction::SINGLE_CONTACT_ACTION_TAG,
       AbstractAction::DATA_MANIPULATION_TAG,